+++ src/main/java/com/buz/buzqb/repository/BrandRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrandRepo extends JpaRepository<Brand, Integer> {
+public interface BrandRepo extends JpaRepository<Brand, Long> {
 
 }
